% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_daynight.R
\name{draw_panel_daynight}
\alias{draw_panel_daynight}
\title{Draw Day/Night Pattern on Panel}
\usage{
draw_panel_daynight(
  data,
  panel_params,
  coord,
  day_fill,
  night_fill,
  sunrise,
  sunset
)
}
\arguments{
\item{data}{The data to be displayed.}

\item{panel_params}{The parameters of the panel.}

\item{coord}{The coordinate system.}

\item{day_fill}{The fill color for daytime rectangles.}

\item{night_fill}{The fill color for nighttime rectangles.}

\item{sunrise}{The hour at which daytime starts.}

\item{sunset}{The hour at which nighttime starts.}
}
\value{
A gList object containing the grobs for the day/night pattern.
}
\description{
Internal function to draw the day/night pattern on the ggplot2 panel.
}
\keyword{internal}
