% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_daynight.R
\docType{data}
\name{GeomDayNight}
\alias{GeomDayNight}
\title{GeomDayNight}
\format{
An object of class \code{GeomDayNight} (inherits from \code{Geom}, \code{ggproto}).
}
\usage{
GeomDayNight
}
\description{
A ggproto object for creating a day/night pattern geom in ggplot2.
}
\details{
This geom creates a pattern along the x-axis of a ggplot2 plot,
distinguishing between daytime and nighttime using rectangles filled
with specified colors.
}
\keyword{internal}
