% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_daynight.R
\name{daynight_table}
\alias{daynight_table}
\title{Create Day/Night Pattern Data}
\usage{
daynight_table(min_datetime, max_datetime, sunrise, sunset)
}
\arguments{
\item{min_datetime}{The starting datetime value (POSIXct format).}

\item{max_datetime}{The ending datetime value (POSIXct format).}

\item{sunrise}{The hour at which daytime starts.}

\item{sunset}{The hour at which nighttime starts.}
}
\value{
A data frame with columns \code{datetime} and \code{daytime},
  where \code{datetime} represents the datetime values and \code{daytime}
  is a logical indicating whether the time is during the day.
}
\description{
Generates a data frame representing daytime and nighttime periods
based on a sequence of datetime values.
}
\keyword{internal}
